requirements.txt
# ============================================================
# ARIA - Automated Responsive Intelligent Assistant
# Requirements - Phase 1 Complete
# ============================================================

# Core AI
google-generativeai==0.5.4

# Voice Input (Speech to Text)
openai-whisper==20231117
sounddevice==0.4.6
numpy==1.26.4
scipy==1.13.0

# Voice Output (Text to Speech)
pyttsx3==2.90

# NLP & Language
langdetect==1.0.9

# Memory & Database
chromadb==0.4.24
sentence-transformers==2.7.0

# File Management
PyPDF2==3.0.1
watchdog==4.0.0

# System Control
psutil==5.9.8
pyautogui==0.9.54
Pillow==10.3.0
pygetwindow==0.0.9

# Email
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0

# Logging & UI
loguru==0.7.2
rich==13.7.1
colorama==0.4.6

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6

# Utilities
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.5

# ── Phase 2 Dependencies ──────────────────────────────────
# WhatsApp
twilio==8.13.0

# Web scraping (Fiverr)
beautifulsoup4==4.12.3
playwright==1.44.0

# Task Scheduling
schedule==1.2.1

# File operations
python-magic==0.4.27

# Phase 2 additions
reportlab==4.1.0
schedule==1.2.1
twilio==8.13.0
beautifulsoup4==4.12.3

# ── Phase 3 Dependencies ──────────────────────────────────
PyPDF2==3.0.1
playwright==1.44.0

# Phase 3
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.31.0
